+++ Test/CoreSDK.Test/Shared/TelemetryClientTest.cs
@@ -483,8 +483,50 @@
 
             Assert.Equal("test name", availability.Name);
             Assert.Equal("test location", availability.RunLocation);
+            Assert.Equal(timestamp, availability.Timestamp);
             Assert.Equal(TimeSpan.FromSeconds(42), availability.Duration);
             Assert.Equal(true, availability.Success);
+        }
+
+        [TestMethod]
+        public void TrackAvailabilityTracksCustomDimensions()
+        {
+            var sentTelemetry = new List<ITelemetry>();
+            TelemetryClient client = this.InitializeTelemetryClient(sentTelemetry);
+
+            var timestamp = DateTimeOffset.Now;
+            var customDimensions = new Dictionary<string,string>()
+                {
+                    ["Blah"] = "yoyo"
+                };
+            
+            client.TrackAvailability("test name", timestamp, TimeSpan.FromSeconds(42), "test location", true, properties: customDimensions);
+
+            var availability = (AvailabilityTelemetry)sentTelemetry.Single();
+
+            Assert.Equal("yoyo", availability.Properties["Blah"]);
+            Assert.Equal(0, availability.Metrics.Count);
+        }
+
+        [TestMethod]
+        public void TrackAvailabilityTracksCustomMetrics()
+        {
+            var sentTelemetry = new List<ITelemetry>();
+            TelemetryClient client = this.InitializeTelemetryClient(sentTelemetry);
+
+            var timestamp = DateTimeOffset.Now;
+
+            var customMetrics = new Dictionary<string, double>()
+            {
+                ["QueueLength"] = 10
+            };
+
+            client.TrackAvailability("test name", timestamp, TimeSpan.FromSeconds(42), "test location", true, metrics: customMetrics);
+
+            var availability = (AvailabilityTelemetry)sentTelemetry.Single();
+
+            Assert.Equal(0, availability.Properties.Count);
+            Assert.Equal(10, availability.Metrics["QueueLength"]);
         }
 
         [TestMethod]
@@ -521,28 +563,14 @@
         public void TrackMethodDoesNotThrowWhenInstrumentationKeyIsEmptyAndNotSendingTheTelemetryItem()
         {
             var channel = new StubTelemetryChannel { ThrowError = true };
-            TelemetryConfiguration.Active = new TelemetryConfiguration
-            {
-                InstrumentationKey = string.Empty,
-                TelemetryChannel = channel
-            };
-
+            TelemetryConfiguration.Active = new TelemetryConfiguration(string.Empty, channel);
             Assert.DoesNotThrow(() => new TelemetryClient().Track(new StubTelemetry()));
         }
 
         [TestMethod]
-        public void ChannelIsInitializedInTrackWhenTelemetryConfigurationIsConstructedViaCtor()
+        public void DefaultChannelInConfigurationIsCreatedByConstructorWhenNotSpecified()
         {
-            TelemetryConfiguration configuration = new TelemetryConfiguration
-            {
-                InstrumentationKey = Guid.NewGuid().ToString()
-            };
-
-            var client = new TelemetryClient(configuration);
-            Assert.Null(configuration.TelemetryChannel);
-
-            client.Track(new StubTelemetry());
-
+            TelemetryConfiguration configuration = new TelemetryConfiguration(Guid.NewGuid().ToString());
             Assert.NotNull(configuration.TelemetryChannel);
         }
 
@@ -555,7 +583,7 @@
 
             ITelemetry sentTelemetry = null;
             var channel = new StubTelemetryChannel { OnSend = telemetry => sentTelemetry = telemetry };
-            var configuration = new TelemetryConfiguration { TelemetryChannel = channel };
+            var configuration = new TelemetryConfiguration(string.Empty, channel);
             var client = new TelemetryClient(configuration);
 
             string environmentKey = Guid.NewGuid().ToString();
@@ -578,7 +606,7 @@
 
             ITelemetry sentTelemetry = null;
             var channel = new StubTelemetryChannel { OnSend = telemetry => sentTelemetry = telemetry };
-            var configuration = new TelemetryConfiguration { TelemetryChannel = channel };
+            var configuration = new TelemetryConfiguration(string.Empty, channel);
             var client = new TelemetryClient(configuration);
 
             string expectedKey = Guid.NewGuid().ToString();
@@ -594,11 +622,7 @@
         [TestMethod]
         public void TrackDoesNotInitializeInstrumentationKeyFromConfigWhenItWasSetExplicitly()
         {
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = new StubTelemetryChannel(),
-                InstrumentationKey = Guid.NewGuid().ToString()
-            };
+            var configuration = new TelemetryConfiguration(Guid.NewGuid().ToString(), new StubTelemetryChannel());
             var client = new TelemetryClient(configuration);
 
             var expectedKey = Guid.NewGuid().ToString();
@@ -613,7 +637,7 @@
         {
             var sentTelemetry = new List<ITelemetry>();
             var channel = new StubTelemetryChannel { OnSend = t => sentTelemetry.Add(t) };
-            var configuration = new TelemetryConfiguration { DisableTelemetry = true , TelemetryChannel = channel };
+            var configuration = new TelemetryConfiguration(string.Empty, channel) { DisableTelemetry = true };
 
             var client = new TelemetryClient(configuration) {};
 
@@ -627,10 +651,9 @@
         {
             var sentTelemetry = new List<ITelemetry>();
             var channel = new StubTelemetryChannel { OnSend = t => sentTelemetry.Add(t) };
-            var configuration = new TelemetryConfiguration { 
-                // no instrumentation key set here
-                TelemetryChannel = channel
-            };
+
+            // No instrumentation key set here.
+            var configuration = new TelemetryConfiguration(string.Empty, channel);
 
             var initializedTelemetry = new List<ITelemetry>();
             var telemetryInitializer = new StubTelemetryInitializer();
@@ -652,11 +675,9 @@
         {
             var sentTelemetry = new List<ITelemetry>();
             var channel = new StubTelemetryChannel { OnSend = t => sentTelemetry.Add(t) };
-            var configuration = new TelemetryConfiguration
-            {
-                // no instrumentation key set here
-                TelemetryChannel = channel
-            };
+
+            // No instrumentation key set here.
+            var configuration = new TelemetryConfiguration(string.Empty, channel);
 
             var initializedTelemetry = new List<ITelemetry>();
             var telemetryInitializer = new StubTelemetryInitializer();
@@ -680,7 +701,7 @@
         [TestMethod]
         public void TrackDoesNotThrowExceptionsDuringTelemetryIntializersInitialize()
         {
-            var configuration = new TelemetryConfiguration { InstrumentationKey = "Test key", TelemetryChannel = new StubTelemetryChannel() };
+            var configuration = new TelemetryConfiguration("Test key", new StubTelemetryChannel());
             var telemetryInitializer = new StubTelemetryInitializer();
             telemetryInitializer.OnInitialize = item => { throw new Exception(); };
             configuration.TelemetryInitializers.Add(telemetryInitializer);
@@ -695,7 +716,7 @@
             {
                 listener.EnableEvents(CoreEventSource.Log, EventLevel.Error);
 
-                var configuration = new TelemetryConfiguration { InstrumentationKey = "Test key", TelemetryChannel = new StubTelemetryChannel() };
+                var configuration = new TelemetryConfiguration("Test key", new StubTelemetryChannel());
                 var telemetryInitializer = new StubTelemetryInitializer();
                 var exceptionMessage = "Test exception message";
                 telemetryInitializer.OnInitialize = item => { throw new Exception(exceptionMessage); };
@@ -720,11 +741,7 @@
                 OnSend = telemetry => sentTelemetry = telemetry,
                 DeveloperMode = false
             };
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = channel,
-                InstrumentationKey = "Test key"
-            };
+            var configuration = new TelemetryConfiguration("Test key", channel);
             var client = new TelemetryClient(configuration);
 
             client.Track(new StubTelemetry());
@@ -741,11 +758,7 @@
                 OnSend = telemetry => sentTelemetry = telemetry,
                 DeveloperMode = true
             };
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = channel,
-                InstrumentationKey = "Test key"
-            };
+            var configuration = new TelemetryConfiguration("Test key", channel);
             var client = new TelemetryClient(configuration);
 
             client.Track(new StubTelemetry());
@@ -757,7 +770,7 @@
         public void TrackDoesNotTryAddingDeveloperModeCustomPropertyWhenTelemetryDoesNotSupportCustomProperties()
         {
             var channel = new StubTelemetryChannel { DeveloperMode = true };
-            var configuration = new TelemetryConfiguration { TelemetryChannel = channel, InstrumentationKey = "Test Key" };
+            var configuration = new TelemetryConfiguration("Test Key", channel);
             var client = new TelemetryClient(configuration);
 
 #pragma warning disable 618
@@ -773,11 +786,7 @@
             {
                 OnSend = telemetry => sentTelemetry = telemetry
             };
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = channel,
-                InstrumentationKey = "Test key"
-            };
+            var configuration = new TelemetryConfiguration("Test key", channel);
             var client = new TelemetryClient(configuration);
 
             client.Track(new StubTelemetry());
@@ -802,12 +811,7 @@
 
             PlatformSingleton.Current = new StubPlatform { OnGetDebugOutput = () => debugOutput };
             var channel = new StubTelemetryChannel { DeveloperMode = true };
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = channel,
-                InstrumentationKey = ""
-
-            };
+            var configuration = new TelemetryConfiguration(string.Empty, channel);
             var client = new TelemetryClient(configuration);
 
             client.Track(new StubTelemetry());
@@ -828,12 +832,7 @@
 
             PlatformSingleton.Current = new StubPlatform { OnGetDebugOutput = () => debugOutput };
             var channel = new StubTelemetryChannel { DeveloperMode = true };
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = channel,
-                InstrumentationKey = "123"
-
-            };
+            var configuration = new TelemetryConfiguration("123", channel);
             var client = new TelemetryClient(configuration);
 
             client.Track(new StubTelemetry());
@@ -851,11 +850,7 @@
             var debugOutput = new StubDebugOutput { OnWriteLine = message => actualMessage = message };
             PlatformSingleton.Current = new StubPlatform { OnGetDebugOutput = () => debugOutput };
             var channel = new StubTelemetryChannel();
-            var configuration = new TelemetryConfiguration
-            {
-                TelemetryChannel = channel,
-                InstrumentationKey = "Test key"
-            };
+            var configuration = new TelemetryConfiguration("Test key", channel);
             var client = new TelemetryClient(configuration);
 
             client.Track(new StubTelemetry());
@@ -866,7 +861,7 @@
         [TestMethod]
         public void TrackCopiesPropertiesFromClientToTelemetry()
         {
-            var configuration = new TelemetryConfiguration { TelemetryChannel = new StubTelemetryChannel() };
+            var configuration = new TelemetryConfiguration(string.Empty, new StubTelemetryChannel());
             var client = new TelemetryClient(configuration);
             client.Context.Properties["TestProperty"] = "TestValue";
             client.Context.InstrumentationKey = "Test Key";
@@ -880,7 +875,7 @@
         [TestMethod]
         public void TrackDoesNotOverwriteTelemetryPropertiesWithClientPropertiesBecauseExplicitlySetValuesTakePrecedence()
         {
-            var configuration = new TelemetryConfiguration { TelemetryChannel = new StubTelemetryChannel() };
+            var configuration = new TelemetryConfiguration(string.Empty, new StubTelemetryChannel());
             var client = new TelemetryClient(configuration);
             client.Context.Properties["TestProperty"] = "ClientValue";
             client.Context.InstrumentationKey = "Test Key";
@@ -900,7 +895,7 @@
             var initializer = new StubTelemetryInitializer();
             initializer.OnInitialize = telemetry => valueInInitializer = ((ISupportProperties)telemetry).Properties[PropertyName];
 
-            var configuration = new TelemetryConfiguration { TelemetryChannel = new StubTelemetryChannel(), TelemetryInitializers = { initializer } };
+            var configuration = new TelemetryConfiguration(string.Empty, new StubTelemetryChannel()) { TelemetryInitializers = { initializer } };
 
             var client = new TelemetryClient(configuration);
             client.Context.Properties[PropertyName] = "ClientValue";
@@ -921,7 +916,7 @@
                     .Version;
             string[] versionParts = new Version(versonStr).ToString().Split('.');
 
-            var configuration = new TelemetryConfiguration { TelemetryChannel = new StubTelemetryChannel(), InstrumentationKey = Guid.NewGuid().ToString() };
+            var configuration = new TelemetryConfiguration(Guid.NewGuid().ToString(), new StubTelemetryChannel());
             var client = new TelemetryClient(configuration);
 
             client.Context.InstrumentationKey = "Test";
@@ -934,7 +929,7 @@
         [TestMethod]
         public void TrackDoesNotOverrideSdkVersion()
         {
-            var configuration = new TelemetryConfiguration { TelemetryChannel = new StubTelemetryChannel(), InstrumentationKey = Guid.NewGuid().ToString() };
+            var configuration = new TelemetryConfiguration(Guid.NewGuid().ToString(), new StubTelemetryChannel());
             var client = new TelemetryClient(configuration);
 
             client.Context.InstrumentationKey = "Test";
@@ -954,8 +949,7 @@
         {
             var sentTelemetry = new List<ITelemetry>();
             var channel = new StubTelemetryChannel { OnSend = t => sentTelemetry.Add(t) };
-            var configuration = new TelemetryConfiguration { InstrumentationKey = "Test key", TelemetryChannel = channel };
-
+            var configuration = new TelemetryConfiguration("Test key", channel);
             var client = new TelemetryClient(configuration);
 
             const int ItemsToGenerate = 100;
@@ -1009,7 +1003,7 @@
             };
 
             // ChuckNorrisTeamUnitTests resource in Prototypes5
-            var config = new TelemetryConfiguration { InstrumentationKey = "fafa4b10-03d3-4bb0-98f4-364f0bdf5df8" };
+            var config = new TelemetryConfiguration("fafa4b10-03d3-4bb0-98f4-364f0bdf5df8");
             var telemetryClient = new TelemetryClient(config);
             telemetryClient.Context.Properties.Add(unicodeString, unicodeString);
             
@@ -1057,7 +1051,7 @@
         private TelemetryClient InitializeTelemetryClient(ICollection<ITelemetry> sentTelemetry)
         {
             var channel = new StubTelemetryChannel { OnSend = t => sentTelemetry.Add(t) };
-            var telemetryConfiguration = new TelemetryConfiguration { InstrumentationKey = Guid.NewGuid().ToString(), TelemetryChannel = channel};
+            var telemetryConfiguration = new TelemetryConfiguration(Guid.NewGuid().ToString(), channel);
             var client = new TelemetryClient(telemetryConfiguration);
             return client;
         }