+++ Test/ServerTelemetryChannel.Test/Shared.Tests/Implementation/TelemetryBufferTest.cs
@@ -71,6 +71,7 @@
             {
                 var buffer = new TelemetryBuffer(new StubTelemetrySerializer(), new StubApplicationLifecycle());
                 Assert.Equal(500, buffer.Capacity);
+                Assert.Equal(1000000, buffer.BacklogSize);
             }
 
             [TestMethod]
@@ -78,14 +79,42 @@
             {
                 var buffer = new TelemetryBuffer(new StubTelemetrySerializer(), new StubApplicationLifecycle());
                 buffer.Capacity = 42;
+                buffer.BacklogSize = 3900;
                 Assert.Equal(42, buffer.Capacity);
+                Assert.Equal(3900, buffer.BacklogSize);
             }
 
             [TestMethod]
-            public void ThrowsArgumentOutOfRangeExceptionWhenNewValueIsLessThanOne()
+            public void ThrowsArgumentExceptionWhenBacklogSizeIsLowerThanCapacity()
+            {
+                var buffer = new TelemetryBuffer(new StubTelemetrySerializer(), new StubApplicationLifecycle());
+                buffer.Capacity = 1111;
+                Assert.Throws<ArgumentException>(() => buffer.BacklogSize = 1110);
+
+                buffer.BacklogSize = 8000;
+                Assert.Throws<ArgumentException>(() => buffer.Capacity = 8001);
+
+            }
+
+            [TestMethod]
+            public void ThrowsArgumentOutOfRangeExceptionWhenNewValueIsLessThanMinimum()
             {
                 var buffer = new TelemetryBuffer(new StubTelemetrySerializer(), new StubApplicationLifecycle());
                 Assert.Throws<ArgumentOutOfRangeException>(() => buffer.Capacity = 0);
+                Assert.Throws<ArgumentOutOfRangeException>(() => buffer.BacklogSize = 0);
+                Assert.Throws<ArgumentOutOfRangeException>(() => buffer.BacklogSize = 1000); // 1001 is minimum anything low would throw.
+
+                bool exceptionThrown = false;
+                try
+                {
+                    buffer.BacklogSize = 1001; // 1001 is valid and should not throw.
+                }
+                catch(Exception)
+                {
+                    exceptionThrown = true;
+                }
+
+                Assert.True(exceptionThrown == false, "No exception should be thrown when trying to set backlogsize to 1001");                
             }
         }
 
@@ -109,6 +138,28 @@
                 buffer.Process(new StubTelemetry());
 
                 Assert.Equal(1, buffer.Count());
+            }
+
+            [TestMethod]
+            public void TelemetryBufferDoesNotGrowBeyondMaxBacklogSize()
+            {
+                //TelemetryBufferWhichDoesNothingOnFlush does not flush items on buffer full,to test ItemDrop scenario.
+                var buffer = new TelemetryBufferWhichDoesNothingOnFlush( new StubTelemetrySerializer(), new StubApplicationLifecycle() );
+                buffer.Capacity = 2;
+                buffer.BacklogSize = 1002;
+
+                buffer.Process(new StubTelemetry());
+
+                // Add more items (1005) to buffer than the max backlog size(1002)
+                for (int i = 0; i < 1005; i++)
+                {
+                    buffer.Process(new StubTelemetry());
+                }
+                                
+                // validate that items are not added after maxunsentbacklogsize is reached.
+                // this also validate that items can still be added after Capacity is reached as it is only a soft limit.
+                int bufferItemCount = buffer.Count();
+                Assert.Equal(1002, bufferItemCount);
             }
 
             [TestMethod]